+++ pybamm/__init__.py
@@ -37,7 +37,15 @@ PARAMETER_PATH = [
 # Utility classes and methods
 #
 from .util import Timer, TimerTime, FuzzyDict
-from .util import root_dir, load_function, rmse, get_infinite_nested_dict, load
+from .util import (
+    root_dir,
+    load_function,
+    rmse,
+    get_infinite_nested_dict,
+    load,
+    is_constant_and_can_evaluate,
+    tree_search,
+)
 from .util import (
     get_parameters_filepath,
     have_jax,
@@ -161,7 +169,8 @@ from .parameters.thermal_parameters import thermal_parameters, ThermalParameters
 from .parameters.lithium_ion_parameters import LithiumIonParameters
 from .parameters.lead_acid_parameters import LeadAcidParameters
 from .parameters.size_distribution_parameters import *
-from .parameters import parameter_sets
+from .parameters.parameter_sets import parameter_sets
+from .parameters_cli import add_parameter, remove_parameter, edit_parameter
 
 #
 # Mesh and Discretisation classes