+++ tests/integration/test_models/standard_model_tests.py
@@ -72,6 +72,11 @@ class StandardModelTest(object):
         self.solver.rtol = 1e-8
         self.solver.atol = 1e-8
 
+        # Somehow removing an equation makes the solver fail at
+        # the low tolerances
+        if isinstance(self.model, pybamm.lithium_ion.NewmanTobias):
+            self.solver.rtol = 1e-7
+
         Crate = abs(
             self.parameter_values["Current function [A]"]
             / self.parameter_values["Nominal cell capacity [A.h]"]