+++ tests/unit/test_util.py
@@ -4,19 +4,10 @@
 import numpy as np
 import os
 import pybamm
-import shutil
 import tempfile
 import unittest
-import importlib
-import subprocess
 from unittest.mock import patch
 from io import StringIO
-
-# Insert .../x/y/z/PyBaMM in sys.path when running this file individually
-import sys
-
-if os.getcwd() not in sys.path:
-    sys.path.insert(0, os.getcwd())
 
 
 class TestUtil(unittest.TestCase):
@@ -25,70 +16,25 @@ class TestUtil(unittest.TestCase):
     """
 
     def test_load_function(self):
-        # Test function load with absolute path
-        abs_test_path = os.path.join(
-            pybamm.root_dir(),
-            "pybamm",
-            "input",
-            "parameters",
-            "lithium_ion",
-            "negative_electrodes",
-            "graphite_Chen2020",
-            "graphite_LGM50_electrolyte_exchange_current_density_Chen2020.py",
-        )
-        func = pybamm.load_function(abs_test_path)
-        self.assertEqual(
-            func,
-            pybamm.input.parameters.lithium_ion.negative_electrodes.graphite_Chen2020.graphite_LGM50_electrolyte_exchange_current_density_Chen2020.graphite_LGM50_electrolyte_exchange_current_density_Chen2020,  # noqa
-        )
+        function = "graphite_LGM50_electrolyte_exchange_current_density_Chen2020"
 
         # Test function load with relative path
         rel_test_path = os.path.join(
-            "pybamm",
-            "input",
-            "parameters",
-            "lithium_ion",
-            "negative_electrodes",
-            "graphite_Chen2020",
-            "graphite_LGM50_electrolyte_exchange_current_density_Chen2020.py",
+            "pybamm", "input", "parameters", "lithium_ion", "Chen2020"
         )
-        func = pybamm.load_function(rel_test_path)
+        func = pybamm.load_function(rel_test_path, function)
         self.assertEqual(
             func,
-            pybamm.input.parameters.lithium_ion.negative_electrodes.graphite_Chen2020.graphite_LGM50_electrolyte_exchange_current_density_Chen2020.graphite_LGM50_electrolyte_exchange_current_density_Chen2020,  # noqa
+            pybamm.input.parameters.lithium_ion.Chen2020.graphite_LGM50_electrolyte_exchange_current_density_Chen2020,  # noqa
         )
 
-        # Test function load for parameters in a directory having "pybamm" in its name
-        # create a new lithium_ion folder in the root PyBaMM directory
-        subprocess.run(["pybamm_edit_parameter", "lithium_ion"])
-
-        # path for a function in the created directory ->
-        # x/y/z/PyBaMM/lithium_ion/negative_electrode/ ....
-        test_path = os.path.join(
-            os.getcwd(),
-            "lithium_ion",
-            "negative_electrodes",
-            "graphite_Chen2020",
-            "graphite_LGM50_electrolyte_exchange_current_density_Chen2020.py",
-        )
-
-        # load the function
-        func = pybamm.load_function(test_path)
-
-        # cannot directly do - lithium_ion.negative_electrodes.graphite_Chen2020 as
-        # lithium_ion is not a python module
-        module_object = importlib.import_module(
-            "lithium_ion.negative_electrodes.graphite_Chen2020.graphite_LGM50_electrolyte_exchange_current_density_Chen2020"  # noqa
-        )
+        # Test function load with absolute path
+        abs_test_path = os.path.join(pybamm.root_dir(), rel_test_path)
+        func = pybamm.load_function(abs_test_path, function)
         self.assertEqual(
             func,
-            getattr(
-                module_object,
-                "graphite_LGM50_electrolyte_exchange_current_density_Chen2020",
-            ),
+            pybamm.input.parameters.lithium_ion.Chen2020.graphite_LGM50_electrolyte_exchange_current_density_Chen2020,  # noqa
         )
-
-        shutil.rmtree("lithium_ion")
 
     def test_rmse(self):
         self.assertEqual(pybamm.rmse(np.ones(5), np.zeros(5)), 1)
@@ -107,6 +53,14 @@ class TestUtil(unittest.TestCase):
         self.assertEqual(d[1][2][3], "x")
         d[4][5] = "y"
         self.assertEqual(d[4][5], "y")
+
+    def test_is_constant_and_can_evaluate(self):
+        symbol = pybamm.PrimaryBroadcast(0, "negative electrode")
+        self.assertEqual(False, pybamm.is_constant_and_can_evaluate(symbol))
+        symbol = pybamm.StateVector(slice(0, 1))
+        self.assertEqual(False, pybamm.is_constant_and_can_evaluate(symbol))
+        symbol = pybamm.Scalar(0)
+        self.assertEqual(True, pybamm.is_constant_and_can_evaluate(symbol))
 
     def test_fuzzy_dict(self):
         d = pybamm.FuzzyDict(